requirements.txt
asyncio
websockets
psycopg2-binary
aiokafka
